bug(#4261): ObjectName as interface
