+++ eo-maven-plugin/src/main/java/org/eolang/maven/MjParse.java
@@ -22,8 +22,8 @@ import org.apache.maven.plugins.annotations.ResolutionScope;
 import org.cactoos.io.InputOf;
 import org.cactoos.iterable.Filtered;
 import org.eolang.parser.EoSyntax;
-import org.eolang.parser.ObjectName;
-import org.eolang.parser.ObjectNameSafe;
+import org.eolang.parser.OnDefault;
+import org.eolang.parser.OnDetailed;
 import org.w3c.dom.Node;
 import org.xembly.Directives;
 import org.xembly.Xembler;
@@ -148,8 +148,8 @@ public final class MjParse extends MjSafe {
             identifier, this.sourcesDir.toPath().relativize(source.toAbsolutePath()), xmir
         );
         final Node document = xmir.inner();
-        final String name = new ObjectNameSafe(
-            new ObjectName(xmir),
+        final String name = new OnDetailed(
+            new OnDefault(xmir),
             e -> MjParse.applyError("mandatory-object-name", e.getMessage(), document)
         ).get();
         if (!name.equals(identifier)) {
